# Eliminates allocations in the binary writer