This patch changes some very hot code paths in the binary writer
to eliminate unnecessary allocations.

In particular:
* Setting the annotations list no longer constructs a temporary
  array of `SymbolToken`s.
* `prepareValue` (which is called for every value in the stream)
  now uses a plain `for` loop instead of an iterator to encode
  each annotation.
* `IonRawBinaryWriter` now uses a custom class called `IntList`
  to store its annotation symbol ID ints instead of a
  `List<Integer>`, which required each int to be boxed.

### Benchmark

This benchmark uses the `ion-java-benchmark-cli` to rewrite an existing ~335MB Ion file.

**Command**
```
java -Xmx8G -jar IJBC.jar write --format ion_binary --ion-flush-period 1024 FILENAME
```

**Before**
```
Benchmark                      Mode  Cnt          Score      Error   Units
Bench.run                        ss   10       5672.757 ±   69.382   ms/op
Bench.run:Heap usage             ss   10       3943.023 ±    0.010      MB
Bench.run:Serialized size        ss   10        335.809                 MB
Bench.run:·gc.alloc.rate         ss   10         33.237 ±    0.360  MB/sec
Bench.run:·gc.alloc.rate.norm    ss   10  216043304.800 ± 5156.126    B/op
Bench.run:·gc.count              ss   10            ≈ 0             counts
```

**After**
```
Benchmark                      Mode  Cnt       Score      Error   Units
Bench.run                        ss   10    5368.811 ±   91.221   ms/op
Bench.run:Heap usage             ss   10    3727.226 ±    1.010      MB
Bench.run:Serialized size        ss   10     335.809                 MB
Bench.run:·gc.alloc.rate         ss   10       0.074 ±    0.001  MB/sec
Bench.run:·gc.alloc.rate.norm    ss   10  457482.400 ± 5146.899    B/op
Bench.run:·gc.count              ss   10         ≈ 0             counts
```

Notice that the allocation rate fell from ~33.237MB/sec to ~0.074MB/sec. The time required also fell by 304ms, a speedup of 5.36%.


By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.